cf/requirements.txt
cryptography
psycopg2-binary
prettytable
flask

